--- a/benchmarks/set_http_meta/scenario.py
+++ b/benchmarks/set_http_meta/scenario.py
@@ -1,78 +1,31 @@
 from collections import defaultdict
+import copy
 from io import BytesIO
 
-import bm
+import bm as bm
 import bm.utils as utils
 
 from ddtrace.contrib.trace_utils import set_http_meta
 
 
 class Config(defaultdict):
-    def __getattribute__(self, item):
+    __header_tag_name = {
+        "User-Agent": "http.user_agent",
+        "REFERER": "http.referer",
+        "Content-Type": "http.content_type",
+        "Etag": "http.etag",
+    }
+
+    def _header_tag_name(self, header_name):
+        return self.__header_tag_name.get(header_name)
+
+    def __getattr__(self, item):
         return self[item]
 
 
-COMMON_DJANO_META = {
-    "PATH": "",
-    "LC_MEASUREMENT": "es_ES.UTF-8",
-    "INVOCATION_ID": "159f1f11bb854f56bd878487c323cadd",
-    "XMODIFIERS": "@im=ibus",
-    "LC_TELEPHONE": "es_ES.UTF-8",
-    "XDG_DATA_DIRS": "/usr/share/ubuntu:/usr/local/share/:/usr/share/:/var/lib/snapd/desktop",
-    "GDMSESSION": "ubuntu",
-    "LC_TIME": "es_ES.UTF-8",
-    "DBUS_SESSION_BUS_ADDRESS": "unix:path=/run/user/1000/bus",
-    "IDE_PROJECT_ROOTS": "//projects/agent",
-    "PS1": "(venv) ",
-    "XDG_CURRENT_DESKTOP": "ubuntu:GNOME",
-    "JOURNAL_STREAM": "8:57546",
-    "SSH_AGENT_PID": "3048",
-    "LC_PAPER": "es_ES.UTF-8",
-    "SESSION_MANAGER": "local/dd:@/tmp/.ICE-unix/3093,unix/dd:/tmp/.ICE-unix/3093",
-    "PWD": "/dd/showcase-django/showcase-django",
-    "MANAGERPID": "2851",
-    "IM_CONFIG_PHASE": "1",
-    "PYCHARM_HOSTED": "1",
-    "GJS_DEBUG_TOPICS": "JS ERROR;JS LOG",
-    "PYTHONPATH": "/dd/testapp_django/py36-39-dj2.2-3.2:/dd/venv:/dd/testapp_flask/py27-39_flask1.1.4",
-    "SHELL": "/usr/bin/zsh",
-    "LC_ADDRESS": "es_ES.UTF-8",
-    "GIO_LAUNCHED_DESKTOP_FILE": "/home/dd/.local/share/applications/jetbrains-pycharm.desktop",
-    "PYENV_ROOT": "/home/dd/.pyenv",
-    "VIRTUAL_ENV": "/dd/venv",
-    "XDG_SESSION_DESKTOP": "ubuntu",
-    "SHLVL": "0",
-    "LC_IDENTIFICATION": "es_ES.UTF-8",
-    "LC_MONETARY": "es_ES.UTF-8",
-    "QT_IM_MODULE": "ibus",
-    "TERM": "xterm-256color",
-    "XDG_CONFIG_DIRS": "/etc/xdg/xdg-ubuntu:/etc/xdg",
-    "LANG": "en_US.UTF-8",
-    "XDG_SESSION_TYPE": "x11",
-    "PYDEVD_LOAD_VALUES_ASYNC": "True",
-    "LC_NAME": "es_ES.UTF-8",
-    "PYCHARM_DISPLAY_PORT": "63342",
-    "XDG_SESSION_CLASS": "user",
-    "_": "/usr/bin/dbus-update-activation-environment",
-    "PYTHONIOENCODING": "UTF-8",
-    "GPG_AGENT_INFO": "/run/user/1000/gnupg/S.gpg-agent:0:1",
-    "DESKTOP_SESSION": "ubuntu",
-    "XDG_MENU_PREFIX": "gnome-",
-    "GIO_LAUNCHED_DESKTOP_FILE_PID": "7043",
-    "QT_ACCESSIBILITY": "1",
-    "WINDOWPATH": "2",
-    "PYTHONDONTWRITEBYTECODE": "1",
-    "LC_NUMERIC": "es_ES.UTF-8",
-    "GJS_DEBUG_OUTPUT": "stderr",
-    "SSH_AUTH_SOCK": "/run/user/1000/keyring/ssh",
-    "PYTHONUNBUFFERED": "1",
-    "GNOME_SHELL_SESSION_MODE": "ubuntu",
-    "XDG_RUNTIME_DIR": "/run/user/1000",
-    "HOME": "/home/dd",
-    "DJANGO_SETTINGS_MODULE": "showcase.settings",
-    "TZ": "UTC",
-    "SERVER_NAME": "localhost",
-    "GATEWAY_INTERFACE": "CGI/1.1",
+PATH = "/test-benchmark/test/1/"
+
+COMMON_DJANGO_META = {
     "SERVER_PORT": "8000",
     "REMOTE_HOST": "",
     "CONTENT_LENGTH": "",
@@ -80,7 +33,7 @@ COMMON_DJANO_META = {
     "SERVER_PROTOCOL": "HTTP/1.1",
     "SERVER_SOFTWARE": "WSGIServer/0.2",
     "REQUEST_METHOD": "GET",
-    "PATH_INFO": "/",
+    "PATH_INFO": PATH,
     "QUERY_STRING": "func=subprocess.run&cmd=%2Fbin%2Fecho+hello",
     "REMOTE_ADDR": "127.0.0.1",
     "CONTENT_TYPE": "text/plain",
@@ -114,36 +67,70 @@ COOKIES = {"csrftoken": "cR8TVoVebF2afssCR16pQeqHcxAlA3867P6zkkUBYDL5Q92kjSGtqpt
 
 DATA_GET = dict(
     method="GET",
+    url="http://localhost:8888{}".format(PATH),
+    status_code=200,
+    status_msg="OK",
+    query=b"key1=value1&key2=value2&token=cR8TVoVebF2afssCR16pQeqHcxAlA3867P6zkkUBYDL5Q92kjSGtqptAry1htdlL",
+    parsed_query={
+        "key1": "value1",
+        "key2": "value2",
+        "token": "cR8TVoVebF2afssCR16pQeqHcxAlA3867P6zkkUBYDL5Q92kjSGtqptAry1htdlL",
+    },
+    request_headers=COMMON_DJANGO_META,
+    response_headers=COMMON_DJANGO_META,
+    retries_remain=0,
+    raw_uri="http://localhost:8888{}?key1=value1&key2=value2&token="
+    "cR8TVoVebF2afssCR16pQeqHcxAlA3867P6zkkUBYDL5Q92kjSGtqptAry1htdlL".format(PATH),
+    request_cookies=COOKIES,
+    request_path_params={"id": 1},
+)
+
+DATA_POST = dict(
+    method="GET",
     url="http://localhost:8888/test-benchmark",
     status_code=200,
     status_msg="OK",
     query=b"key1=value1&key2=value2",
     parsed_query={"key1": "value1", "key2": "value2"},
-    request_headers=COMMON_DJANO_META,
-    response_headers=COMMON_DJANO_META,
+    request_headers=COMMON_DJANGO_META,
+    response_headers=COMMON_DJANGO_META,
     retries_remain=0,
     raw_uri="http://127.0.0.1:8081/test/1/?key1=value1&key2=value2",
     request_cookies=COOKIES,
     request_path_params={"id": 1},
-    request_body=None,  # type: Optional[Union[str, Dict[str, List[str]]]]
+    request_body=None,
 )
 
 
 class SetHttpMeta(bm.Scenario):
-    nspans = bm.var(type=int)
     allenabled = bm.var_bool()
+    useragentkey = bm.var_bool()
 
     def run(self):
         # run scenario to also set tags on spans
-        span = utils.gen_span(str(self))
-        config_all_enabled = Config(lambda: True)
-        config_all_disabled = Config(lambda: False)
+        if self.allenabled:
+            config = Config(lambda: True)
+        else:
+            config = Config(lambda: False)
 
-        def _(loops):
+        data = copy.deepcopy(DATA_GET)
+        if self.useragentkey:
+            del data["request_headers"]["HTTP_USER_AGENT"]
+
+        def bm(loops):
             for _ in range(loops):
-                if self.allenabled:
-                    set_http_meta(span, config_all_enabled, **DATA_GET)
-                else:
-                    set_http_meta(span, config_all_disabled, **DATA_GET)
+                span = utils.gen_span(str("test"))
+                span._local_root = utils.gen_span(str("root"))
+                set_http_meta(span, config, **data)
 
-        yield _
+        yield bm
+
+
+#
+# if __name__ == "__main__":
+#     # run scenario to also set tags on spans
+#     span = utils.gen_span(str("test"))
+#     span._local_root = utils.gen_span(str("root"))
+#     config_all_enabled = Config(lambda: True)
+#     set_http_meta(span, config_all_enabled, **DATA_GET)
+#     print(span.get_tags())