--- a/benchmarks/set_http_meta/scenario.py
+++ b/benchmarks/set_http_meta/scenario.py
@@ -43,8 +43,6 @@ COMMON_DJANGO_META = {
     "HTTP_SEC_CH_UA": '"Chromium";v="92", " Not A;Brand";v="99", "Google Chrome";v="92"',
     "HTTP_SEC_CH_UA_MOBILE": "?0",
     "HTTP_UPGRADE_INSECURE_REQUESTS": "1",
-    "HTTP_USER_AGENT": "Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 "
-    "(KHTML, like Gecko) Chrome/92.0.4515.159 Safari/537.36",
     "HTTP_ACCEPT": "text/html,application/xhtml+xml,application/xml;q=0.9,image/avif,image/webp,"
     "image/apng,*/*;q=0.8,application/signed-exchange;v=b3;q=0.9",
     "HTTP_SEC_FETCH_SITE": "none",
@@ -105,6 +103,7 @@ DATA_POST = dict(
 class SetHttpMeta(bm.Scenario):
     allenabled = bm.var_bool()
     useragentkey = bm.var_bool()
+    useragentvariants = bm.var_bool()
 
     def run(self):
         # run scenario to also set tags on spans
@@ -114,23 +113,20 @@ class SetHttpMeta(bm.Scenario):
             config = Config(lambda: False)
 
         data = copy.deepcopy(DATA_GET)
+
+        user_agent_keys = []
         if self.useragentkey:
-            del data["request_headers"]["HTTP_USER_AGENT"]
+            user_agent_keys = ["HTTP_USER_AGENT"]
+            if self.useragentvariants:
+                user_agent_keys = ["HTTP_USER_AGENT", "USER_AGENT", "UserAgent", "User-Agent", "user-agent"]
 
         def bm(loops):
             for _ in range(loops):
-                span = utils.gen_span(str("test"))
-                span._local_root = utils.gen_span(str("root"))
-                set_http_meta(span, config, **data)
+                for user_agent_key in user_agent_keys:
+                    data["request_headers"][user_agent_key] = "Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 " \
+                                                              "(KHTML, like Gecko) Chrome/92.0.4515.159 Safari/537.36"
+                    span = utils.gen_span(str("test"))
+                    span._local_root = utils.gen_span(str("root"))
+                    set_http_meta(span, config, **data)
 
         yield bm
-
-
-#
-# if __name__ == "__main__":
-#     # run scenario to also set tags on spans
-#     span = utils.gen_span(str("test"))
-#     span._local_root = utils.gen_span(str("root"))
-#     config_all_enabled = Config(lambda: True)
-#     set_http_meta(span, config_all_enabled, **DATA_GET)
-#     print(span.get_tags())