--- a/benchmarks/set_http_meta/scenario.py
+++ b/benchmarks/set_http_meta/scenario.py
@@ -124,6 +124,7 @@ class SetHttpMeta(bm.Scenario):
 
         def bm(loops):
             for _ in range(loops):
-                set_http_meta(span, config, **data)
+                for _ in range(10000):
+                    set_http_meta(span, config, **data)
 
         yield bm