--- a/benchmarks/span/scenario.py
+++ b/benchmarks/span/scenario.py
@@ -1,5 +1,5 @@
 import bm
-import utils
+import bm.utils as utils
 
 
 class Span(bm.Scenario):