## Description
Add benchmark tests to `ddtrace.contrib.trace_utils.set_http_meta`


Results between branch #3987  and 1.4.0rc1
```
+-------------------------------+--------------------------------------------------------------------------------------+-------------------------------------------------------------------------------------------------------+
| Benchmark                     | /artifacts/05a38bf3-a628-46d2-93f4-1c397cd3ac85/set_http_meta/1.4.0rc1//results.json | /artifacts/05a38bf3-a628-46d2-93f4-1c397cd3ac85/set_http_meta/1.3.0rc2.dev113+gbcf523c5//results.json |
+===============================+======================================================================================+=======================================================================================================+
| sethttpmeta-start             | 13.5 us                                                                              | 13.9 us: 1.03x slower                                                                                 |
+-------------------------------+--------------------------------------------------------------------------------------+-------------------------------------------------------------------------------------------------------+
| sethttpmeta-all-enabled       | 29.3 us                                                                              | 28.2 us: 1.04x faster                                                                                 |
+-------------------------------+--------------------------------------------------------------------------------------+-------------------------------------------------------------------------------------------------------+
| sethttpmeta-all-disabled      | 12.4 us                                                                              | 13.8 us: 1.12x slower                                                                                 |
+-------------------------------+--------------------------------------------------------------------------------------+-------------------------------------------------------------------------------------------------------+
| sethttpmeta-no-useragentkey   | 43.8 ns                                                                              | 43.2 ns: 1.01x faster                                                                                 |
+-------------------------------+--------------------------------------------------------------------------------------+-------------------------------------------------------------------------------------------------------+
| sethttpmeta-useragentvariants | 151 us                                                                               | 155 us: 1.02x slower                                                                                  |
+-------------------------------+--------------------------------------------------------------------------------------+-------------------------------------------------------------------------------------------------------+
| Geometric mean                | (ref)                                                                                | 1.02x slower                                                                                          |
+-------------------------------+--------------------------------------------------------------------------------------+-------------------------------------------------------------------------------------------------------+
```

## Checklist
- [x] Title must conform to [conventional commit](https://github.com/conventional-changelog/commitlint/tree/master/%40commitlint/config-conventional).
- [ ] Add additional sections for `feat` and `fix` pull requests.
- [ ] Ensure tests are passing for affected code.
- [ ] [Library documentation](https://github.com/DataDog/dd-trace-py/tree/1.x/docs) and/or [Datadog's documentation site](https://github.com/DataDog/documentation/) is updated. Link to doc PR in description.

## Reviewer Checklist
- [ ] Title is accurate.
- [ ] Description motivates each change.
- [ ] No unnecessary changes were introduced in this PR.
- [ ] PR cannot be broken up into smaller PRs.
- [ ] Avoid breaking [API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces) changes unless absolutely necessary.
- [ ] Tests provided or description of manual testing performed is included in the code or PR.
- [ ] Release note has been added for fixes and features, or else `changelog/no-changelog` label added.
- [ ] All relevant GitHub issues are correctly linked.
- [ ] Backports are identified and tagged with Mergifyio.
- [ ] Add to milestone.